split up subscription file
