+++ src/api/interfaces/chain.rs
@@ -1,0 +1,60 @@
+/*
+   Copyright 2019 Supercomputing Systems AG
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+	   http://www.apache.org/licenses/LICENSE-2.0
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+*/
+
+use crate::{
+	api::{error::Error, Api, ApiResult, TransactionStatus},
+	rpc::{HandleSubscription, Subscribe},
+	utils, XtStatus,
+};
+use ac_compose_macros::rpc_params;
+pub use ac_node_api::{events::EventDetails, StaticEvent};
+use ac_node_api::{DispatchError, Events};
+use ac_primitives::{ExtrinsicParams, FrameSystemConfig};
+use log::*;
+use serde::de::DeserializeOwned;
+use sp_core::storage::StorageChangeSet;
+
+pub type TransactionSubscriptionFor<Client, Hash> =
+	<Client as Subscribe>::Subscription<TransactionStatus<Hash, Hash>>;
+
+pub trait ChainSubscription<Client, Hash>
+where
+	Client: Subscribe,
+	Hash: DeserializeOwned,
+{
+	type Header: DeserializeOwned;
+
+	fn subscribe_finalized_heads(&self) -> ApiResult<Client::Subscription<Self::Header>>;
+}
+
+impl<Signer, Client, Params, Runtime> NodeSubscription<Client, Runtime::Hash>
+	for Api<Signer, Client, Params, Runtime>
+where
+	Client: Subscribe,
+	Params: ExtrinsicParams<Runtime::Index, Runtime::Hash>,
+	Runtime: FrameSystemConfig,
+	Runtime::Header: DeserializeOwned,
+{
+	type Header = Runtime::Header;
+
+	fn subscribe_finalized_heads(&self) -> ApiResult<Client::Subscription<Self::Header>> {
+		debug!("subscribing to finalized heads");
+		self.client()
+			.subscribe(
+				"chain_subscribeFinalizedHeads",
+				rpc_params![],
+				"chain_unsubscribeFinalizedHeads",
+			)
+			.map_err(|e| e.into())
+	}
+}
