+++ src/api/interfaces/mod.rs
@@ -12,13 +12,14 @@
 */
 
 pub use self::{
-	frame_system::*, pallet_balances::*, pallet_transaction_payment::*, state::*,
-	submit_extrinisc::*, subscription::*,
+	author::*, frame_system::*, pallet_balances::*, pallet_transaction_payment::*, state::*,
+	subscription::*,
 };
 
+pub mod author;
+pub mod chain;
 pub mod frame_system;
 pub mod pallet_balances;
 pub mod pallet_transaction_payment;
 pub mod state;
-pub mod submit_extrinisc;
 pub mod subscription;
